Rename `go-fuzz` to `ctb-test-case-generator`
