+++ go.work
@@ -19,7 +19,7 @@ use (
 	./op-proposer
 	./op-service
 	./proxyd
-	./packages/contracts-bedrock/go-fuzz
+	./packages/contracts-bedrock/test-case-generator
 )
 
 replace github.com/ethereum/go-ethereum v1.10.26 => github.com/ethereum-optimism/op-geth v0.0.0-20221205191237-0678a130d790
